+++ src/main/resources/release_notes.md
@@ -53,6 +53,7 @@
 * Add a Linux-friendly default theme, called “Gnome.”
 * Add Kotlin support to codebase
 * Reset progress bar when action is finished (#1955)
+* Upgrade Gradle to version 8.5
 
 ## Translations